requirements.txt
discord.py
requests
beautifulsoup4
apscheduler
google-play-scraper 